+++ docs/changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Generalize all scripts
 - Translate system identifiers in scg-editor
 - View sc-links identifiers
 - Load sc-links identifiers from gwf